refactor: single market widget
