+++ lib/ui/views/home/presentation/widgets/single_market_widget.dart
@@ -1,0 +1,55 @@
+import 'package:flutter/material.dart';
+import 'package:gap/gap.dart';
+import 'package:stock/ui/shared/styles/colors.dart';
+import 'package:stock/ui/shared/styles/fonts.dart';
+
+class SingleMarketWidget extends StatelessWidget {
+  const SingleMarketWidget({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {
+        print("yes");
+      },
+      child: Container(
+        margin: const EdgeInsets.only(bottom: 15),
+        padding: const EdgeInsets.symmetric(
+          vertical: 20,
+          horizontal: 20,
+        ),
+        decoration: BoxDecoration(
+          color: Colors.transparent,
+          borderRadius: BorderRadius.circular(7),
+          border: Border.all(color: Colors.grey[300]!),
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            CustomText(
+              'Agilent Technologies Inc.',
+              color: Colors.black,
+              fontWeight: FontWeight.w600,
+            ),
+            const Gap(5),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                SmallText(
+                  'Nvidia',
+                  color: AppColors.grey,
+                ),
+                SmallText(
+                  'USD',
+                  color: AppColors.black,
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
